Add Makefile to build xpi and install local app
The default target just builds an XPI extension packages that can be
normally installed (provided xpinstall.signatures.required is false).

There are furthermore the local-install and local-uninstall targets for
placing the native application at the right location or removing it
again.

Finally, the release target allows to update the manifest with the new
version, commits that and tags the result.

All this is derived from my TB extensions such as copypatch

[Felix] Fix extension name in .gitignore

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>
Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>