feat: use real ssoUrl when requesting PRT SSO cookie
When requesting a PRT SSO cookie, the ssoUrl must be provided. In the
past, we just used a static default, as only the non-emptyness of the
parameter is checked. However, future authentication backends might also
check that value, hence already provide it.

The Microsoft Edge implementation also passes the value down to the PRT
SSO cookie aquisition function.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>