Eliminate pylint findings
We don't want to change the module name, so suppress "invalid-name" at
the beginning and then re-enable it to keep the code itself covered.

The 'object' argument name is not our invention but coming from pydbus.
Suppress that.

It is unclear why pylint complains about using a method as callback for
signal_fired while the code works fine. Ignore this warning.

Yes, GRAPH_SCOPES is a mutable default, but also ignore this for now.

Docstrings can come later, ignore that as well.

Finally, hook up the check with CI so that we don't regress again.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>