fix: show useful information on account reload
In 94f1afd all logging was moved to a function. However, this function
does not process varargs. By that, the information of the current logged
in account was not shown in the terminal. This patch fixes this by
showing the email address of the current account instead of the JS
object.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>