feat: handle broker state changes
This patch decouples the state "enabled" from "broker-online". By that,
we can start in the enabled state and wait for the broker to become
online (fixes starting FF right after login, where the broker is not yet
online). Also, restarts of the broker (e.g. due to OOM) are now handled
explicitly in the frontend code. Here, the icon and the tooltip text
inform the user about the current state.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>