feat: strictly follow ms-oapxbc protocol
The ms-oapxbc OAuth2 protocol extension specifies to send the PRT token
directly as header (not encapsulated in a cookie). While all naming of
the broker functions suggestes that the data should be encoded in a
cookie, the publicly available documentation clearly states to use a
header.

In short, the header works as well and also lets us drop the more
complicated cookie injection. By that, we switch.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>