fix: load extension on startup of firefox
With our recent migration to Manifest v3, the extension is only loaded
on the first event (i.e. the first hit of the ms login page). This is
problematic as the first hit on that login page then will not have the
PRT SSO cookie injected. Further, it makes it harder to debug issues as
the profile picture is not shown in the app menu until this page is hit.

Instead, we now also register a trigger for the onStartup event to
initialize the extension when the browser is started. This migration is
based on [1] and required some refactoring.

[1] https://extensionworkshop.com/documentation/develop/manifest-v3-migration-guide/

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>