chore: deploy firefox addon via gh-pages
The idea is to deploy the update manifest for Mozilla Firefox to github
pages. On new releases, we create the release and build the (unsigned)
xpi. Then, we upload that to AMO and wait for it to be signed. Once
signed, we download the signed xpi and attach that to the release. Then,
we update the update manifest and push that to main (which updates it on
gh-pages as well). After that, all Firefox instances that have updates
for this extension enabled will automatically download and updatethe
extension.

closes: #5

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>