fix: always terminate host component when browser is terminated
On chrome and chromium, the native part is not reliably terminated when
the browser is terminated. As we also cannot reliably send a message to
the host part on termination of the browser, there is no direct way to
let the host script known. In theory, the script should terminate once
stdio is closed, but in practice this is not always the case, leading to
orphaned processes.

To fix this, we use prctl to register a PDEATHSIG handler.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>