fix: reset api token before requesting new
Before requesting a new API token, we need to clear the variable that holds
the current one, as this state is used as a synchronization point to
wait for the token response.

This fixes an issue, where the profile picture could not be loaded,
because an outdated token was used.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>