ci: attach build provenance attestation to artifacts
To attest that our artifacts were actually built in the CI, we add
provenance attestations. We add these for all non-released artifacts, as
well as for the signed Firefox artifact we get back from mozilla. For
the Chrome / Chromium artifacts, we don't add attestations, as we do not
get these artifacts back from the store during CI execution.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>