ci: fix incorrect release tag by getting information from GH
With the existing git checkout configuration, the current tag was not
correctly detected by the versioning logic in the Makefile. This led to
incorrect version numbers of the debian package.

Instead, we now explicitly set the RELEASE_TAG to the information we get
from GitHub, when building the release package in CI.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>