fix: remove extension-id placeholder in installed chromium manifest
The native messaging manifest of chrome and chromium contains two
allowed extension ids: A placeholder for a local (non-signed) install
and the offical extension id (signed version). When installing globally,
the placeholder for the local version needs to be removed. Due to a typo
in the Makefile, this happened only for chrome, but not for chromium.

Reported-by: Dr. Johann Pfefferl <johann.pfefferl@siemens.com>
Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>