chore: check if broker is running before starting
In case the identity broker is not registered as systemd service, we
cannot start it. This however is also not needed, as long as the service
is visible on the bus.

This patch changes the logic to check for a running broker by first
checking if it is running and only if not starting it.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>