refactor: decouple UI from internal state
Correctly maintaining the state both internally as well as with the UI
became more and more complex, especially after adding the menu.

To mitigate this in a future proof way, we decouple the state from the
UI representation. By that, the logic blocks are much easier to read and
understand. Also other developers can work on the UI stuff more easily.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>