feat: show plugin versions in menu
When updating the browser plugin via the app stores, the host version
might get out of sync with the web extension version.

To notify the user about this, we add an endpoint (API call) to query
version information of the host tooling. In case it does not match
we show both versions in the UI. Later on, we can define version
boundaries and show a more prominent warning. For now, this is not
needed as we are 100% compatible with the first released version.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>