fix(workaround): wait for broker DBus path to become visible
The broker implements DBus activation. However, it appears on the bus
before it can fully be used (i.e. before the
/com/microsoft/identity/broker1 becomes visible). By that, the buffered
API calls are dispatched too early and return with an UnknownObject
error. This affects both the introspection and API calls.

To work around this, we poll for the path to show up before performing
any other broker communication. Once the broker disappears, we clear our
internal state to ensure the polling happens again before the next
invocation.

This logic makes it possible to run the broker only when needed and also
solves issues around the initial startup of the broker via the web
extension.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>