feat: remain logged in during broker outages
Previously, we precisely tracked the broker state and logged the user
out in case the broker disappeared. This however made it impossible to
auto-disable the broker after some time and only re-enable it when
needed again.

Now, we keep the user logged in once the frontend communicated once with
the broker. When needed again, we just call into the backend and let it
handle the startup, waiting and errors.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>