chore: make license detectable by licensee
The licensee tool is used by github (and others) to detect the name,
version and attribution of a repositories license. In case of github,
this data is then shown in the summary page, making it easily available
to users. As we did not provide a LICENSE or copying file, but just
stated the license in the README.md, the tooling was not able to detect
the license with a sufficiently high confidence to show this as MPL-2.0.

We copy the license file from the LICENSES dir to the repo root, as
neither licensee nor reuse lint supports symlinks. The change does not
change the license itself in any form. It is just about making it
accessible to tooling.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>