chore: change Python interpreter on global installs
When installing the host components system-wide, we set the path to the
Python interpreter to a concrete one (/usr/bin/python3). By that, the
script is not affected by settings of a local user session (e.g. that
use a default venv for development).

As this path might vary across distributions, we make it configurable in
"python3_bin".

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>