ci: run build workflow on external pull requests
Currently we don't run the build workflow on external pull requests.
This makes it tricky to detect build and style issues on these requests.

As we already require explicit approval for all external contributors
before running the action, it is fine to also enable the workflow.
Further, I carefully checked the configuration of the repo with the
poutine tool for possible bypasses.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>