feat: indicate broker online state in UI
In case the connection to the broker is lost, an indication of that
situation should be given to the user. While this is not an error by
itself (given that the broker is dbus-activated), it still provides
valuable information for debugging. We now indicate this by adding a red
circle around the avatar in this case, as well as stating the current
status in the menu (when clicking on the icon).

Proposed-by: Jan Kiszka <jan.kiszka@siemens.com>
Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>