call broker only if really needed
Whenever the internal code state changes, e.g. because of a login or
broker-becomes-available event, we acquire a new PrtSsoCookie (chrome)
or re-inject the http rules. This is not needed and creates additional
overhead on the potentially slow communication via DBus. In most cases,
the ruleset is not changed anyways, but just some UI metadata needs to
be updated.

We mitigate this by adding a ui_only parameter to the state update
handler which conditionally skips the rule and PRT update part.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>