fix: load dummy avatar prior to real one
In case an error happens between "login" (assignment of the account) and
loading the profile picture, we don't have any profile picture set. This
is problematic as downstream functions require a valid image data
object. By that, the background worker was in an invalid state which
required an extension restart to recover.

To fix this, we load the dummy image upfront and then wait for the actual
image to be available.

Reported-by: Florian Bezdeka <florian.bezdeka@siemens.com>
Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>