feat: cache login state across browser restarts
Recently we got reports about missing SSO in case the browser is
not already running when opening the Entra SSO page. This happens
because the login flow takes some time (in case of broker dbus
activation a couple of seconds) and only after it is completed we enable
the request filtering and try to acquire PRTs.

We now change this by caching the last logged-in user in the local
browser storage. By that, the extension can already request PRTs before
the login completed, as we already have the needed account information
to request the PRT. A nice side effect of that change is that we now
track the disabled / enabled state across browser restarts as well.

As this change requires the use of the local storage API, we also need
to grant permission in the manifests.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>