fix: set correct accept type when requesting avatar
When requesting the avatar picture, we set the Content-Type header
instead of the Accept header. This is wrong, as the Content-Type header
is (usually) a response header.

As the server anyways always responded with an image/jpeg type, this
did not have any user visible effect. Anyways, we should clean this up,
which we do here.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>