refactor(popup): avoid use of innerHTML where not needed
When setting the content of some UI elements, we are just interested in
setting text values, but no HTML. By that, we can use the safe
.innerText method to protect ourselves against HTML injection.

This is not a security fix, as the data we set there is provided by
trusted sources. However, it is good practice and also avoids a linter
warning.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>