chore(ui): report broken connection to host tooling
In the field we saw multiple errors were the browser was not able to
communicate with the host part. This usually is the case if a flatpak
version of the browser is used or the host tooling simply was not
installed. We now report this error via the tooltip of the disabled
extension.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>