fix: do not send image data to broker
We previously used the same object to store the active account as
returned from the broker and internal metadata (e.g. the avatar image).
By that, we also sent the image data to the broker via the NM backend
and finally via DBus.

We fix this by introducing a new account class, which internally stores
the account as returned by the broker and provide members to access the
needed fields. We further add logic to store and transmit only the
serializable parts.

No user impact.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>