fix: do not send image data to broker
We previously used the same object to store the active account as
returned from the broker and internal metadata (e.g. the avatar image).
By that, we also sent the image data to the broker via the NM backend
and finally via DBus.

We fix this by splitting the object into a raw part that contains the
account as returned by the broker and other members that we internally
use.

No user impact.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>