feat: support transparent token refresh on well-known apps
On Firefox, we use the WebRequest API to inject the PRT SSO Cookie into
the requests hitting the login provider. As the `WebRequest` API only operates
on allowed URLs, applications that perform an automated token refresh in the
background (e.g. the Teams PWA) also need to be listed. Otherwise the
transparent re-login will fail (it misses the `PRT SSO Cookie`) and a manual
re-login will be required. The same applies to the declarativeNetRequest API
on chromium based browsers. As we now filter based on request type, we drop
the nearly redundant request access header type filter.

To require only the absolute minimum set of permissions, we do this on a
case-by-case decision instead of whitelisting all URLs. We further
update the privacy statement to include a reasoning. This commit just
prepares the infrastructure. No apps are added.

As we now rely on optional_host_permissions, we need to lift the minimal
required Firefox version to 128. This is anyways the current ESR
version, so we do not expect older versions in the field.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>