feat: account for dynamically changed permissions
Host permissions are not automatically granted on app update. By that,
users need to either grant them manually via the extensions page, or
when visiting an SSO enabled page, by right-clicking the app icon.
In this case, we need to reload the extension internal filters to enable
the new ruleset.

As this process is quite error prone (e.g. users forget to allow the SSO
url), we add logic to notify the user via an entry in the extension
menu. This later on can be extended to show which apps are enabled.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>