feat: allow users to enable background sso on arbitrary pages
Maintaing the allow-list does not scale, as in theory any page can use
the Entra SSO login provider. As we do not want to require access to all
requests on any page for privacy reasons, we add the https://*/* wildcard
filter to optional host permissions. By that, the user can allow the SSO
on any page by explicitly granting access via the extension menu.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>