feat: allow to configure background SSO domains via policy
Prior to this patch, the user was responsible to enable the background
SSO on a per-domain basis. We now implement the managed storage API to
allow configuring the enabled domain via policies. For technical
reasons, a one-time user interaction is still required, as the policy
cannot grant the permissions by itself: When a policy update is
available, the menu shows a hint with a button that needs to be clicked
by the user. The policy update then only overrides settings provided in
the policy. All other user settings remain.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>