fix(Firefox): register onBeforeSendHeaders listener only once
Previously the listener might have been installed multiple times,
despite we never saw this in practice. That's likely because the same
function was called, which simply got overwritten. Unfortunately the
spec is not precise in this regard.

We now follow the recommendation to unregister the listener before we
re-register. This is especially important, as the arguments might have
been changed in the meantime (e.g. more URLs got enabled). As the
unregister is a noop in case the listener is not registered, we don't
check upfront.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>