ci: do not auto publish on CWS on releases
The CWS is very strict about changes in the manifest file (especially
permission related changes). The need for the changes often explicitly
needs to be explained before being allowed to publish (and starting the
CWS review process). By automatically trying to publish, the CI step
fails in this case, marking the CI of the git tag as failed.

As this is somehow expected behavior, we no longer automatically
publish, but just upload to the CWS. Then, the maintainers can inspect
the initial report, answer the questions and publish manually. We
further document this process in the CONTRIBUTING guide.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>