refactor: prepare project for platform specific code
While improving the support for Thunderbird, it became clear, that we
need to separate the platform specific code from the generic to be able
to externally set which browser we are running on. This anyways makes
sense, as we currently trigger release warnings regarding unsupported
interfaces (in inactive code path).

As a preparation, we reorganize the repository and introduce modules to
organize the code.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>