refactor: move native messaging to separate module
The native messaging part heavily relied on global state. This made it
impossible to move platform specific code into separate modules, as
these cannot access the globals. Further, the RPC logic was partially
modeled as RPC and partially modeled event driven, making it hard to
maintain.

We clean this up by moving the native messaging parts to a new broker
module (and class). This wraps the native messaging with an async
RPC interface that can easily be used from various contexts. For that,
we introduce a RPC queue helper class, as we need to match the requests
and responses. The broker is semantically identical to the DBus
interface and just handles the data passing.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>