refactor: move token injection logic to platform
On Firefox and Chrome the logic to inject the PRT SSO cookie is
fundamentally different. Previously, both code paths were in the code
and a runtime dispatch was used to select the correct one. However, this
already created warnings regarding unsupported APIs in the addon review.

Recently, platform specific modules were introduced, so we move the code
there. In addition, this reduces the amount of global state we rely on.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>