chore: only runtime differentiate between Firefox and Thunderbird
Firefox and Thunderbird currently use the same platform. However, there
are slight UI optimizations needed for TB which currently rely on a
runtime dispatch. We need to keep it, while we don't have a separate
platform for Thunderbird. However, we can dispatch all other browser
parts to the platform logic.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>