feat: make Thunderbird standalone platform
The Thunderbird version unfortunately needs a slightly different
manifest. Also, in the future we might want to fine-tune things a bit.

For now, we just build a dedicated .xpi, but do not include it in the
release process. It is still unclear, if the auto distribution logic we
currently use is compatible with the Thunderbird app store (and a first
time manual upload is needed anyways).

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>