fix: notify state change only after broker is initialized
As the notify_state_change accesses the broker object to check if the
connection is running, we also need to instantiate the broker first.
Currently, the extension fails to load, as the broker object is null at
this point in time. This obvious error looks like a rebase artifact,
which was introduced in da8b99c.

Fixes: da8b99c ("refactor: move native messaging to separate module")
Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>