ci(thunderbird): build TB extension with mozilla tooling
The Thunderbird version of the extension does not need to be signed by
Mozilla, as self-distributed extensions for Thunderbird do not require
this. However, we still want to build it with the same tooling we use
for the signed version. By that, also the filenames match our glob
patterns for the release upload.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>