chore(thunderbird): set application id
As told by Mozilla [1], we should use a distinct application id for the
thunderbird version. This also gives us the chance to select a new
application id in the now preferred format (@string instead of email) [2].

To keep the id similar to the previous one, we now use
"@linux-entra-sso.tb". Note, that this id cannot be changed later on
without creating a "new" extension entry.

[1] https://discourse.mozilla.org/t/listing-webextension-on-both-amo-and-addons-thunderbird-net/145536/3
[2] https://developer.mozilla.org/en-US/docs/Mozilla/Add-ons/WebExtensions/manifest.json/browser_specific_settings

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>