feat(ui): unify text scaling across browsers
Chrome and Firefox use different default font sizes for the UI menu. By
that, the text was shown super-small on Chrome (especially on high-dpi
screens), making it hard to interact with the UI. We fix that by
explicitly defining the body font size and viewport. We further use
variables for that, so later on we can easily adapt if needed.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>