feat(ui): indicate immutable bg sso state with text as well
In case the background SSO is managed via policy for the current URL,
the only indicator if it is enabled or disabled is the color of the dot
next to the URL. This is tricky to on small screens and basically
impossible for color blind people.

To fix this usability issue, we display the current state next to the
"Background SSO" line in text form as well. Further, we provide a
tooltip, that this cannot be changed because it is set via policy.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>