feat(ui): do not show background SSO toggle on SSO url
When visiting the login provider url, we currently offer the option to
disable the background SSO. This makes no sense, because this url must
be enabled to even allow foreground SSO. If not granted, the extension
is not functional (which is noted in a error message in the UI anyways).

We now additionally check if we are on the SSO url and hide the
background SSO menu in this case.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>