feat: add support for multiple accounts
Currently only the first account that is reported back by the broker can
be used to inject SSO tokens. This breaks for users that have multiple
accounts for different purposes (e.g. user vs. admin accounts).

We now list all accounts in the UI and let the user select it. This
assignment is also stable across restarts of the extension or the
browser.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>