fix: only register handlers if operational
We already distinguish the state is_operational (we have everything to
perform SSO) and is_active, which just denotes that the user wants to
use SSO (this does not mean all necessary pre-conditions are true).

For the request handlers, we need to check if we are operational.
This bug usually was not critical, as later on we got proper account
data and re-initialized the request handlers.

Fixes: febc859 ("refactor(accounts): move account handling logic ...")
Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>