chore(ui): improve formatting of long account entries
In case an account has a very long email address or name, not the whole
available space was used. Instead, it was broken up into many lines. We
now ensure we always use the whole available space and only break if
really needed. By that, we also avoid pixel-precise coupling of the
width of the popup and internal entries.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>