docs: improve readability of README
The README describes the installation of the host tooling on a
per-browser level. As the steps are almost identical, we group them to
reduce the amount of duplicated text. We further improve the readability
by re-ordering some aspects to stress that the extension consists of
both native and browser parts.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>