chore(thunderbird): show extension in account setup dialog
When setting up an account via the "Account Setup" dialog, the extension
currently is hidden. Unfortunately this dialog has no dedicated space
assigned, so we need to add the "default" space to show it there.

Especially in this dialog it is critical to show the extension so the
user can select the correct SSO account and check if the SSO is enabled.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>