feat: track compliance state of device
The device compliant state is an important information for the user, as
it affects which services the user can access. Getting the compliance
data is a two-staged process: First extract the device id from a broker
issued access token. Then, we use that to query the current compliance
state from the Graph API.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>