chore(ui): set font family of body element
By explicitly setting a font-family for the body element, we align the
looking of the extension across browsers. This further helps to align
the rendering of our unicode indicator dot, which got improved in 2e3f220
but looks quite big on some systems without pinning the font.

We further decide to use a sans-serif font (as chromium anyways used)
for a cleaner rendering of small texts.

Xref: #101
Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>