chore(ui): limit string length of device and url
It can happen that device names (as FQDN) and the current tab URL (used
in background SSO UI element) are very long, leading to UI glitches
where the alignment between the right and left column is broken.

We fix this by limiting the string length to a reasonable value and add
a tooltip with the full string if it is overly long.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>